Prevent overflow in ray tracing example

+++ ray_tracing/src/Pixmap.h
@@ -14,7 +14,7 @@
 
 class Pixmap {
 public:
-    using PixelType = Vec3<uint8_t>;
+    using PixelType = Vec3<float>;
 
     Pixmap(uint32_t width, uint32_t height);
